Move CalypsoCard and StorageCard from data to domain
