+++ src/control/app/src/main/kotlin/org/calypsonet/keyple/demo/control/domain/managers/CalypsoCardManager.kt
@@ -1,33 +1,22 @@
-/* ******************************************************************************
- * Copyright (c) 2021 Calypso Networks Association https://calypsonet.org/
- *
- * See the NOTICE file(s) distributed with this work for additional information
- * regarding copyright ownership.
- *
- * This program and the accompanying materials are made available under the
- * terms of the BSD 3-Clause License which is available at
- * https://opensource.org/licenses/BSD-3-Clause.
- *
- * SPDX-License-Identifier: BSD-3-Clause
- ****************************************************************************** */
-package org.calypsonet.keyple.demo.control.data
+package org.calypsonet.keyple.demo.control.domain.managers
 
-import java.time.LocalDateTime
 import org.calypsonet.keyple.demo.common.constants.CardConstants
 import org.calypsonet.keyple.demo.common.model.ContractStructure
 import org.calypsonet.keyple.demo.common.model.EventStructure
 import org.calypsonet.keyple.demo.common.model.type.PriorityCode
 import org.calypsonet.keyple.demo.common.model.type.VersionNumber
-import org.calypsonet.keyple.demo.common.parsers.*
+import org.calypsonet.keyple.demo.common.parsers.ContractStructureParser
+import org.calypsonet.keyple.demo.common.parsers.EnvironmentHolderStructureParser
+import org.calypsonet.keyple.demo.common.parsers.EventStructureParser
+import org.calypsonet.keyple.demo.control.domain.mappers.ContractMapper
+import org.calypsonet.keyple.demo.control.domain.mappers.ValidationMapper
 import org.calypsonet.keyple.demo.control.domain.model.AppSettings
+import org.calypsonet.keyple.demo.control.domain.model.AuthenticationMode
+import org.calypsonet.keyple.demo.control.domain.model.CardReaderResponse
 import org.calypsonet.keyple.demo.control.domain.model.Contract
 import org.calypsonet.keyple.demo.control.domain.model.Location
 import org.calypsonet.keyple.demo.control.domain.model.Status
 import org.calypsonet.keyple.demo.control.domain.model.Validation
-import org.calypsonet.keyple.demo.control.data.mappers.ContractMapper
-import org.calypsonet.keyple.demo.control.data.mappers.ValidationMapper
-import org.calypsonet.keyple.demo.control.domain.model.AuthenticationMode
-import org.calypsonet.keyple.demo.control.domain.model.CardReaderResponse
 import org.eclipse.keyple.card.calypso.CalypsoExtensionService
 import org.eclipse.keypop.calypso.card.WriteAccessLevel
 import org.eclipse.keypop.calypso.card.card.CalypsoCard
@@ -39,8 +28,9 @@ import org.eclipse.keypop.calypso.card.transaction.TransactionManager
 import org.eclipse.keypop.reader.CardReader
 import org.eclipse.keypop.reader.ChannelControl
 import timber.log.Timber
+import java.time.LocalDateTime
 
-class CalypsoCardImpl {
+class CalypsoCardManager {
 
   fun executeControlProcedure(
       controlDateTime: LocalDateTime,
@@ -99,7 +89,8 @@ class CalypsoCardImpl {
       // <Abort Secure Session if any>
       if (env.envVersionNumber != VersionNumber.CURRENT_VERSION) {
         if (cardTransaction is SecureRegularModeTransactionManager ||
-            cardTransaction is SecurePkiModeTransactionManager) {
+            cardTransaction is SecurePkiModeTransactionManager
+        ) {
           cardTransaction.prepareCancelSecureSession().processCommands(ChannelControl.CLOSE_AFTER)
         }
         throw EnvironmentException("wrong version number")
@@ -109,7 +100,8 @@ class CalypsoCardImpl {
       // <Abort Secure Session if any>
       if (env.envEndDate.getDate().isBefore(controlDateTime.toLocalDate())) {
         if (cardTransaction is SecureRegularModeTransactionManager ||
-            cardTransaction is SecurePkiModeTransactionManager) {
+            cardTransaction is SecurePkiModeTransactionManager
+        ) {
           cardTransaction.prepareCancelSecureSession().processCommands(ChannelControl.CLOSE_AFTER)
         }
         throw EnvironmentException("End date expired")
@@ -131,7 +123,8 @@ class CalypsoCardImpl {
       val eventVersionNumber = event.eventVersionNumber
       if (eventVersionNumber != VersionNumber.CURRENT_VERSION) {
         if (cardTransaction is SecureRegularModeTransactionManager ||
-            cardTransaction is SecurePkiModeTransactionManager) {
+            cardTransaction is SecurePkiModeTransactionManager
+        ) {
           cardTransaction.prepareCancelSecureSession().processCommands(ChannelControl.CLOSE_AFTER)
         }
         if (eventVersionNumber == VersionNumber.UNDEFINED) {
@@ -272,12 +265,13 @@ class CalypsoCardImpl {
         }
       }
 
-      Timber.i("Control procedure result: STATUS_OK")
+      Timber.Forest.i("Control procedure result: STATUS_OK")
       status = Status.TICKETS_FOUND
 
       // Step 20 - If a session is open, Close the session
       if (cardTransaction is SecureRegularModeTransactionManager ||
-          cardTransaction is SecurePkiModeTransactionManager) {
+          cardTransaction is SecurePkiModeTransactionManager
+      ) {
         cardTransaction.prepareCloseSecureSession().processCommands(ChannelControl.CLOSE_AFTER)
       }
 
@@ -291,10 +285,11 @@ class CalypsoCardImpl {
           status = status,
           authenticationMode = authenticationMode,
           lastValidationsList = validationList,
-          titlesList = displayedContract)
+          titlesList = displayedContract
+      )
     } catch (e: Exception) {
       errorMessage = e.message
-      Timber.e(e)
+      Timber.Forest.e(e)
       when (e) {
         is EnvironmentException -> {
           errorMessage = "Environment error: $errorMessage"
@@ -316,7 +311,8 @@ class CalypsoCardImpl {
         authenticationMode = authenticationMode,
         titlesList = arrayListOf(),
         errorTitle = errorTitle,
-        errorMessage = errorMessage)
+        errorMessage = errorMessage
+    )
   }
 
   /**
